pyproject.toml
[build-system]
requires = ["setuptools>=68.0", "wheel"]
build-backend = "setuptools.backends._legacy:_Backend"

[project]
name = "aegis-deploy"
version = "0.1.0"
description = "Production deployment harness for the Aegis medical image de-identification pipeline — MAP packaging, Argo orchestration, and AWS infrastructure."
readme = "README.md"
license = {text = "Apache-2.0"}
requires-python = ">=3.10"
authors = [
    {name = "Aegis Team"},
]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Healthcare Industry",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Topic :: Scientific/Engineering :: Medical Science Apps.",
]

dependencies = [
    # Aegis core library (de-identification engine)
    "monai-aegis @ git+https://github.com/lakshmi-mahabaleshwara/aegis.git#subdirectory=monai_aegis",
    # MONAI Deploy Application Package SDK
    "monai-deploy-app-sdk>=3.0.0",
    # AWS
    "boto3>=1.34",
    "s3fs>=2024.2",
    # Database (Identity Vault)
    "sqlalchemy>=2.0",
    "psycopg2-binary>=2.9",
    # Config & utilities
    "pyyaml>=6.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0",
    "pytest-cov>=5.0",
    "moto[s3]>=5.0",
    "black>=24.0",
    "flake8>=7.0",
    "mypy>=1.8",
]

[project.scripts]
aegis-deploy = "aegis_deploy.cli:main"

[tool.setuptools.packages.find]
include = ["aegis_deploy*"]

[tool.pytest.ini_options]
testpaths = ["tests"]

[tool.black]
line-length = 100
target-version = ["py310"]

[tool.mypy]
python_version = "3.10"
warn_return_any = true
warn_unused_configs = true
